Changed type from Integer to Long in Order entity id field

+++ src/main/java/pl/rarytas/rarytas_restaurantside/controller/rest/OrderRestController.java
@@ -86,7 +86,7 @@ public class OrderRestController {
     }
 
     @GetMapping("/id/{id}")
-    public ResponseEntity<?> getById(@PathVariable Integer id) {
+    public ResponseEntity<?> getById(@PathVariable Long id) {
         Optional<?> result = orderService.findById(id);
         return result.map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
